meetapp/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==2.2.4
django-adminactions==1.6.0
django-filter==2.4.0
djangorestframework==3.12.4
envparse==0.2.0
httplib2==0.19.1
idna==2.10
lxml==4.6.3
Pillow==6.1.0
psycopg2-binary==2.8.3
pyparsing==2.4.7
pytz==2019.2
requests==2.25.1
selenium==3.141.0
six==1.12.0
soupsieve==2.2.1
sqlparse==0.3.0
undetected-chromedriver==2.2.7
unicodecsv==0.14.1
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
